+++ test/flex_sub_themes_test.dart
@@ -52,7 +52,7 @@ void main() {
         'EXPECT equal to AppBarTheme with shape RoundedRectangleBorder()', () {
       expect(
         FlexSubThemes.appBarTheme(),
-        equals(const AppBarTheme(shape: RoundedRectangleBorder())),
+        equals(const AppBarThemeData(shape: RoundedRectangleBorder())),
       );
     });
   });
@@ -4532,7 +4532,7 @@ void main() {
           useMaterial3: useM3, // false
         ).toString(minLevel: DiagnosticLevel.fine),
         equalsIgnoringHashCodes(
-          InputDecorationTheme(
+          InputDecorationThemeData(
             labelStyle: WidgetStateTextStyle.resolveWith((
               Set<WidgetState> states,
             ) {
@@ -4745,7 +4745,7 @@ void main() {
               ? usedFillColor.darken(kInputDecoratorLightBgDarken)
               : usedFillColor.lighten(kInputDecoratorDarkBgLighten);
 
-      final InputDecorationTheme m = FlexSubThemes.inputDecorationTheme(
+      final InputDecorationThemeData m = FlexSubThemes.inputDecorationTheme(
         colorScheme: colorScheme,
         tintedDisabled: tintDisable,
         tintedInteractions: tintInteract,
@@ -4755,7 +4755,7 @@ void main() {
       expect(
         m.toString(minLevel: DiagnosticLevel.fine),
         equalsIgnoringHashCodes(
-          InputDecorationTheme(
+          InputDecorationThemeData(
             labelStyle: WidgetStateTextStyle.resolveWith((
               Set<WidgetState> states,
             ) {
@@ -4991,7 +4991,7 @@ void main() {
           useMaterial3: useM3,
         ).toString(minLevel: DiagnosticLevel.fine),
         equalsIgnoringHashCodes(
-          InputDecorationTheme(
+          InputDecorationThemeData(
             labelStyle: WidgetStateTextStyle.resolveWith((
               Set<WidgetState> states,
             ) {
@@ -5226,7 +5226,7 @@ void main() {
           useMaterial3: useM3,
         ).toString(minLevel: DiagnosticLevel.fine),
         equalsIgnoringHashCodes(
-          InputDecorationTheme(
+          InputDecorationThemeData(
             labelStyle: WidgetStateTextStyle.resolveWith((
               Set<WidgetState> states,
             ) {
@@ -5462,7 +5462,7 @@ void main() {
           useMaterial3: useM3,
         ).toString(minLevel: DiagnosticLevel.fine),
         equalsIgnoringHashCodes(
-          InputDecorationTheme(
+          InputDecorationThemeData(
             labelStyle: WidgetStateTextStyle.resolveWith((
               Set<WidgetState> states,
             ) {
@@ -5759,7 +5759,7 @@ void main() {
           useMaterial3: false,
         ).toString(minLevel: DiagnosticLevel.fine),
         equalsIgnoringHashCodes(
-          InputDecorationTheme(
+          InputDecorationThemeData(
             labelStyle: WidgetStateTextStyle.resolveWith((
               Set<WidgetState> states,
             ) {
@@ -6004,7 +6004,7 @@ void main() {
       );
       //
       // Outline M3 case
-      InputDecorationTheme m = FlexSubThemes.inputDecorationTheme(
+      InputDecorationThemeData m = FlexSubThemes.inputDecorationTheme(
         colorScheme: colorScheme,
         borderType: FlexInputBorderType.outline,
         backgroundAlpha: 0xAA,
@@ -11846,7 +11846,7 @@ void main() {
           colorScheme: colorScheme,
           backgroundColor: const Color(0xFFDDDDDD),
           backgroundSchemeColor: SchemeColor.tertiary,
-          inputDecorationTheme: const InputDecorationTheme(),
+          inputDecorationTheme: const InputDecorationThemeData(),
           useInputDecoratorTheme: true,
           useMaterial3: false,
         ),
@@ -11882,7 +11882,7 @@ void main() {
         FlexSubThemes.timePickerTheme(
           colorScheme: colorScheme,
           backgroundColor: const Color(0xFFDDDDDD),
-          inputDecorationTheme: const InputDecorationTheme(filled: true),
+          inputDecorationTheme: const InputDecorationThemeData(filled: true),
           useInputDecoratorTheme: true,
           useMaterial3: false,
         ),
@@ -11922,7 +11922,7 @@ void main() {
         FlexSubThemes.timePickerTheme(
           colorScheme: colorScheme,
           backgroundSchemeColor: SchemeColor.tertiary,
-          inputDecorationTheme: const InputDecorationTheme(filled: true),
+          inputDecorationTheme: const InputDecorationThemeData(filled: true),
           elementRadius: 12,
           useInputDecoratorTheme: true,
           useMaterial3: false,
@@ -11959,7 +11959,7 @@ void main() {
       );
       final TimePickerThemeData m = FlexSubThemes.timePickerTheme(
         colorScheme: colorScheme,
-        inputDecorationTheme: const InputDecorationTheme(filled: true),
+        inputDecorationTheme: const InputDecorationThemeData(filled: true),
         useInputDecoratorTheme: true,
         useMaterial3: true,
       );